file with prompts
A first batch of musical prompts. Chat GPT does not want to create 10000 at once, so I do it step by step. However, I think that some are actually duplicates, and many sound the same.

My goal is to reach 1000 by Monday, but maybe we need another, more flexible mechanism to create these prompts. My idea: prompts that consist of 3 parts that are each randomized, so we just have to save the parts and create a combination of these parts on the spot. E.g. "British, break-up, rock" -> A  song from a British rock band that you associate with the hurt feeling of a break-up. But we can discuss this.